requirements.txt
discord.py
peewee
colored
# watchdog
onnxruntime
pillow
peewee